Remove Patch; Use standard query params instead
Also: move code around & create a separate DecisionMatrix file
