+++ explorer/Explorer.jsdom.test.tsx
@@ -22,21 +22,21 @@ describe(Explorer, () => {
 
     it("each grapher has its own set of URL params/options are preserved even when the grapher changes", () => {
         const explorer = element.instance() as Explorer
-        expect(explorer.patchObject.tab).toBeUndefined()
+        expect(explorer.queryParams.tab).toBeUndefined()
 
         explorer.onChangeChoice("Gas Radio")("All GHGs (CO₂eq)")
 
         if (explorer.grapher) explorer.grapher.tab = GrapherTabOption.table
         else throw Error("where's the grapher?")
-        expect(explorer.patchObject.tab).toEqual("table")
+        expect(explorer.queryParams.tab).toEqual("table")
 
         explorer.onChangeChoice("Gas Radio")("CO₂")
-        expect(explorer.patchObject.tab).toBeUndefined()
+        expect(explorer.queryParams.tab).toBeUndefined()
     })
 
     it("recovers country selection from URL params", () => {
         const element = mount(
-            SampleExplorer({ uriEncodedPatch: "selection~Ireland" })
+            SampleExplorer({ queryStr: "?selection=Ireland" })
         )
         const explorer = element.instance() as Explorer
         expect(explorer.selection.selectedEntityNames).toEqual(["Ireland"])
