+++ explorer/tsconfig.json
@@ -9,7 +9,6 @@
         { "path": "../grapher" },
         { "path": "../gridLang" },
         { "path": "../coreTable" },
-        { "path": "../patch" },
         { "path": "../urls" }
     ]
 }
