--- patch/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "extends": "../devTools/tsconfigs/tsconfig.base.json",
-    "compilerOptions": {
-        "outDir": "../itsJustJavascript/patch",
-        "rootDir": "."
-    }
-}
