+++ tsconfig.json
@@ -13,7 +13,6 @@
         { "path": "./explorer" },
         { "path": "./coreTable" },
         { "path": "./site" },
-        { "path": "./patch" },
         { "path": "./settings" },
         { "path": "./adminSiteClient" },
         { "path": "./adminSiteServer" },
